HV-1001 cleaned up checkstyle violations. moved NotEmpty validators to separate package. updated Bootsrapping test as NotEmpty is not using NotNull anymore  corrected population of NotEmpty Validators in ConstraintHelper.  started to change Array Validators to separate out arrays of primitives as initial approach is allowing to put this annotation on non compatible field types.
