+++ engine/src/main/java/org/hibernate/validator/internal/constraintvalidators/hv/empty/NotEmptyCharSequenceValidator.java
@@ -4,7 +4,7 @@
  * License: Apache License, Version 2.0
  * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
  */
-package org.hibernate.validator.internal.constraintvalidators.hv;
+package org.hibernate.validator.internal.constraintvalidators.hv.empty;
 
 /**
  * Check that a character sequence's (e.g. string) length is greater than zero.
