+++ engine/src/main/java/org/hibernate/validator/internal/constraintvalidators/hv/empty/NotEmptyIterableValidator.java
@@ -4,7 +4,7 @@
  * License: Apache License, Version 2.0
  * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
  */
-package org.hibernate.validator.internal.constraintvalidators.hv;
+package org.hibernate.validator.internal.constraintvalidators.hv.empty;
 
 /**
  * Check that a iterable length is greater than zero.
