+++ engine/src/test/java/org/hibernate/validator/test/internal/constraintvalidators/hv/EmptyArrayValidatorTest.java
@@ -12,7 +12,7 @@ import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
 
 import org.hibernate.validator.constraints.NotEmpty;
-import org.hibernate.validator.internal.constraintvalidators.hv.NotEmptyArrayValidator;
+import org.hibernate.validator.internal.constraintvalidators.hv.empty.NotEmptyArrayValidator;
 import org.testng.annotations.Test;
 
 import javax.validation.ConstraintViolation;
